Pipeline/requirements.txt
python-dotenv
sqlalchemy
pandas
requests
--no-binary :all: pyodbc 